feat: agregar modelos, controladores, migraciones y seeders para clientes y pedidos

+++ Código/Backend/Ejemplos/Laravel/ETS_Ejemplo_01/database/migrations/2026_07_13_231129_create_clientes_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('clientes', function (Blueprint $table) {
-            $table->bigInteger('ci')->primary();
+            $table->unsignedInteger('ci')->primary();
             $table->timestamps();
             $table->string('nombre',50);
             $table->string('apellido',50);
